requirements.txt
aioambient==2025.2.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiokafka==0.12.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
bidict==0.23.1
brotlicffi==1.0.9.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.7.0
neo4j==5.26.0
packaging==25.0
pip==25.2
propcache==0.3.1
pycares==4.10.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
Pygments==2.20.0
PySocks==1.7.1
python-dotenv~=1.2.2
python-engineio==4.12.2
python-multipart==0.0.22
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
starlette==0.47.3
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
yarl==1.22.0

nats-py~=2.14.0